requirements.txt
transformers==4.20.1
torch==1.12.0
numpy==1.21.6
pillow==7.1.2